Server and Client code compiles again.
git-svn-id: svn://10.0.50.55/legends/trunk@1261 67207eb2-9daa-044d-9a34-b8567af74898

+++ code/src/GEL2/Grid/SubGrid2D_Class.h
@@ -364,12 +364,12 @@ public:
 	// - -------------------------------------------------------------------------------------- - //
 	// Get the position, returning the dead value if over //
 	inline const szt DeadIndex( const int _x, const int _y, const szt DeadValue = SZT_MAX ) const {
-		if ( _x >= Width() )
+		if ( _x >= (int)Width() )
 			return DeadValue;
 		else if ( _x < 0 )
 			return DeadValue;
 			
-		if ( _y >= Height() )
+		if ( _y >= (int)Height() )
 			return DeadValue;
 		else if ( _y < 0 )
 			return DeadValue;
